Getting a distroted image for some reason.
All the three colors do show, but are out of place.
Would be interesting to know where did I let it loose.

One more thing that I realized is taht the objects in the final picture
are leaning towards left, due to which, one circle cuts off, and is
shown on the rigth side. "The problem is something is making the things d
rift to the left, as of now".